+++ CHANGELOG.md
@@ -14,11 +14,12 @@
 ## Changed
 - Residual plot in `pyvela-plot` script
 - Made GP noise marginalization the default in `SPNTA`
-- Move `info_dict` from `pyvela_script` to `SPNTA`
+- Move `info_dict`, `save_new_parfile`, `save_resids` from `pyvela_script` to `SPNTA`
 ## Fixed
 - Julia example script `run_example.jl`
 - In `_gls_lnlike_serial`, return  -inf if `Σinv` is not positive definite.
 ## Removed
+- Saving maximum-posterior par file in `pyvela` script
 
 # 0.0.9
 ## Added